requirements.txt
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
defusedxml==0.5.0
Django==2.2
django-filter==2.1.0
django-settings-export==1.2.1
djangorestframework==3.9.2
djangorestframework-jsonp==1.0.2
djangorestframework-simplejwt==4.3.0
djangorestframework-xml==1.4.0
djangorestframework-yaml==1.0.3
kombu==4.5.0
PyJWT==1.7.1
pymongo==3.7.2
python-memcached==1.59
pytz==2019.1
PyYAML==5.1
six==1.12.0
sqlparse==0.3.0
vine==1.3.0
